refactor exchange
