+++ .travis.yml
@@ -17,7 +17,7 @@ script:
   - golint -set_exit_status ./...
   - gofmt -l . | exit $(wc -l)
   - go test -cover -race -count 100 ./...
-  - go test -v -run nothing -bench=. ./...
+  - go test -v -run nothing -bench=. -benchmem ./...
 
 notifications:
   slack:
